+++ inc/geometries/ds_curv_t.cu
@@ -44,7 +44,7 @@ int main(int argc, char * argv[])
     std::cout << "# Constructing Grid..."<<std::endl;
     dg::geo::CurvilinearProductGrid3d g3d(flux, n, Nx, Ny,Nz, dg::NEU);
     std::cout << "# Constructing Fieldlines..."<<std::endl;
-    dg::geo::DS<dg::aProductGeometry3d, dg::IDMatrix, dg::DMatrix, dg::DVec> ds( mag, g3d, dg::NEU, dg::PER, dg::geo::FullLimiter(), 1e-8, mx, my, -1, method);
+    dg::geo::DS<dg::aProductGeometry3d, dg::IDMatrix, dg::DVec> ds( mag, g3d, dg::NEU, dg::PER, dg::geo::FullLimiter(), 1e-8, mx, my, -1, method);
 
     t.toc();
     std::cout << "# Construction took "<<t.diff()<<"s\n";