+++ inc/geometries/ds_guenter_t.cu
@@ -38,7 +38,7 @@ int main( )
     ////////////////////////////////initialze fields /////////////////////
     const dg::CylindricalGrid3d g3d( R_0 - a, R_0+a, -a, a, 0, 2.*M_PI, n, Nx, Ny, Nz, dg::NEU, dg::NEU, dg::PER);
     const dg::geo::TokamakMagneticField mag = dg::geo::createGuenterField(R_0, I_0);
-    dg::geo::DS<dg::aProductGeometry3d, dg::IDMatrix, dg::DMatrix, dg::DVec> ds(
+    dg::geo::DS<dg::aProductGeometry3d, dg::IDMatrix, dg::DVec> ds(
         mag, g3d, dg::NEU, dg::NEU, dg::geo::FullLimiter(),
         1e-8, mx, my, -1, method);